Bug 1229437 followup 2 - Fix another sign-compare error in RubyColumn::Iterator on CLOSED TREE.

+++ layout/generic/RubyUtils.h
@@ -163,7 +163,8 @@ struct MOZ_STACK_CLASS RubyColumn
       , mIndex(aIndex)
     {
       MOZ_ASSERT(aIndex == -1 ||
-                 (aIndex >= 0 && aIndex <= aColumn.mTextFrames.Length()));
+                 (aIndex >= 0 &&
+                  aIndex <= int32_t(aColumn.mTextFrames.Length())));
       SkipUntilExistingFrame();
     }
     friend struct RubyColumn; // for the constructor
